separate settings code into components
