+++ src/components/Settings.vue
@@ -53,1012 +53,44 @@
       </div>
 
       <div>
-        <div v-if="activeTab === 'core'" class="space-y-6">
-          <h3 class="text-xl font-semibold mb-4 text-blue-400">
-            Core API Configuration
-          </h3>
-          <fieldset
-            class="fieldset bg-gray-900/90 border-blue-500/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">API Keys & Providers</legend>
-            <div class="grid grid-cols-1 md:grid-cols-2 gap-4 p-2">
-              <div>
-                <label for="ai-provider" class="block mb-1 text-sm"
-                  >AI Provider *</label
-                >
-                <select
-                  id="ai-provider"
-                  v-model="currentSettings.aiProvider"
-                  class="select select-bordered w-full focus:select-primary"
-                >
-                  <option value="openai">OpenAI</option>
-                  <option value="openrouter">OpenRouter</option>
-                  <option value="ollama">Ollama (Local)</option>
-                  <option value="lm-studio">LM Studio (Local)</option>
-                </select>
-              </div>
-              <div>
-                <label for="stt-provider" class="block mb-1 text-sm"
-                  >Speech-to-Text Provider *</label
-                >
-                <select
-                  id="stt-provider"
-                  v-model="currentSettings.sttProvider"
-                  class="select select-bordered w-full focus:select-primary"
-                >
-                  <option value="openai">OpenAI (gpt-4o-transcribe)</option>
-                  <option value="groq">Groq (whisper-large-v3)</option>
-                </select>
-              </div>
-              <div>
-                <label for="openai-key" class="block mb-1 text-sm"
-                  >OpenAI API Key *</label
-                >
-                <input
-                  id="openai-key"
-                  type="password"
-                  v-model="currentSettings.VITE_OPENAI_API_KEY"
-                  class="input focus:outline-none w-full"
-                  autocomplete="new-password"
-                  placeholder="sk-..."
-                />
-                <p class="text-xs text-gray-400 mt-1">
-                  Required for TTS/STT/embeddings regardless of AI provider.
-                </p>
-              </div>
-              <div v-if="currentSettings.aiProvider === 'openrouter'">
-                <label for="openrouter-key" class="block mb-1 text-sm"
-                  >OpenRouter API Key *</label
-                >
-                <input
-                  id="openrouter-key"
-                  type="password"
-                  v-model="currentSettings.VITE_OPENROUTER_API_KEY"
-                  class="input focus:outline-none w-full"
-                  autocomplete="new-password"
-                  placeholder="sk-or-v1-..."
-                />
-                <p class="text-xs text-gray-400 mt-1">
-                  Required for chat models when using OpenRouter.
-                </p>
-              </div>
-              <div v-if="currentSettings.aiProvider === 'ollama'">
-                <label for="ollama-url" class="block mb-1 text-sm"
-                  >Ollama Base URL *</label
-                >
-                <input
-                  id="ollama-url"
-                  type="text"
-                  v-model="currentSettings.ollamaBaseUrl"
-                  class="input focus:outline-none w-full"
-                  placeholder="http://localhost:11434"
-                />
-                <p class="text-xs text-gray-400 mt-1">
-                  URL where your Ollama server is running.
-                </p>
-              </div>
-              <div v-if="currentSettings.aiProvider === 'lm-studio'">
-                <label for="lmstudio-url" class="block mb-1 text-sm"
-                  >LM Studio Base URL *</label
-                >
-                <input
-                  id="lmstudio-url"
-                  type="text"
-                  v-model="currentSettings.lmStudioBaseUrl"
-                  class="input focus:outline-none w-full"
-                  placeholder="http://localhost:1234"
-                />
-                <p class="text-xs text-gray-400 mt-1">
-                  URL where your LM Studio server is running.
-                </p>
-              </div>
-              <div v-if="currentSettings.sttProvider === 'groq'">
-                <label for="groq-key" class="block mb-1 text-sm"
-                  >Groq API Key (for STT) *</label
-                >
-                <input
-                  id="groq-key"
-                  type="password"
-                  v-model="currentSettings.VITE_GROQ_API_KEY"
-                  class="input focus:outline-none w-full"
-                  autocomplete="new-password"
-                  placeholder="gsk_..."
-                />
-                <p class="text-xs text-gray-400 mt-1">
-                  Required only if Groq STT is selected above.
-                </p>
-              </div>
-            </div>
-          </fieldset>
-        </div>
+        <CoreSettingsTab 
+          v-if="activeTab === 'core'"
+          :current-settings="currentSettings"
+        />
 
-        <div v-if="activeTab === 'assistant'" class="space-y-6">
-          <h3 class="text-xl font-semibold mb-4 text-green-400">
-            Assistant Configuration
-          </h3>
+        <AssistantSettingsTab 
+          v-if="activeTab === 'assistant'"
+          :current-settings="currentSettings"
+          :available-models="availableModelsForSelect"
+          :available-tools="availableToolsForSelect"
+          :is-refreshing-models="isRefreshingModels"
+          :is-tool-configured="isToolConfigured"
+          @refresh-models="refreshModels"
+          @reset-system-prompt="resetSystemPrompt"
+        />
 
-          <fieldset
-            class="fieldset bg-gray-900/90 border-green-500/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">Model & Behavior</legend>
-            <div class="space-y-4 p-2">
-              <div>
-                <label
-                  for="assistant-model"
-                  class="block mb-1 text-sm flex items-center"
-                  >Assistant Model *
-                  <div
-                    class="tooltip tooltip-right"
-                    data-tip="The core model used for generating responses. A more powerful model will provide better results but will be more expensive."
-                  >
-                    <img :src="infoIcon" class="size-4 ml-1" /></div
-                ></label>
-                <select
-                  id="assistant-model"
-                  v-model="currentSettings.assistantModel"
-                  class="select select-bordered w-full focus:select-primary"
-                >
-                  <option disabled value="">Select a model</option>
-                  <option
-                    v-if="
-                      conversationStore.availableModels.length === 0 &&
-                      settingsStore.coreOpenAISettingsValid
-                    "
-                    value=""
-                  >
-                    Loading models...
-                  </option>
-                  <option
-                    v-for="model in availableModelsForSelect"
-                    :key="model.id"
-                    :value="model.id"
-                  >
-                    {{ model.id }}
-                  </option>
-                </select>
-                <button
-                  type="button"
-                  @click="refreshModels"
-                  :disabled="isRefreshingModels"
-                  class="btn btn-sm btn-outline btn-primary mt-2"
-                >
-                  <span v-if="isRefreshingModels" class="loading loading-spinner loading-xs mr-2"></span>
-                  {{ isRefreshingModels ? 'Loading...' : 'Refresh Models' }}
-                </button>
-                <p
-                  v-if="
-                    !settingsStore.coreOpenAISettingsValid &&
-                    ((currentSettings.aiProvider === 'openai' &&
-                      currentSettings.VITE_OPENAI_API_KEY) ||
-                      (currentSettings.aiProvider === 'openrouter' &&
-                        currentSettings.VITE_OPENROUTER_API_KEY) ||
-                      (currentSettings.aiProvider === 'ollama' &&
-                        currentSettings.ollamaBaseUrl) ||
-                      (currentSettings.aiProvider === 'lm-studio' &&
-                        currentSettings.lmStudioBaseUrl)) &&
-                    conversationStore.availableModels.length === 0
-                  "
-                  class="text-xs text-warning mt-1"
-                >
-                  {{
-                    getProviderDisplayName(currentSettings.aiProvider)
-                  }}
-                  API key/configuration needs to be validated (Save & Test) to load models.
-                </p>
-              </div>
+        <HotkeysTab 
+          v-if="activeTab === 'hotkeys'"
+          :current-settings="currentSettings"
+          :is-recording-hotkey-for="isRecordingHotkeyFor"
+          @start-recording-hotkey="startRecordingHotkey"
+          @clear-hotkey="clearHotkey"
+        />
 
-              <div>
-                <div class="flex justify-between items-center mb-1">
-                  <label for="assistant-system-prompt" class="block text-sm"
-                    >Assistant System Prompt</label
-                  >
-                  <button
-                    type="button"
-                    @click="resetSystemPrompt"
-                    class="btn btn-xs btn-ghost"
-                  >
-                    Reset
-                  </button>
-                </div>
-                <textarea
-                  id="assistant-system-prompt"
-                  v-model="currentSettings.assistantSystemPrompt"
-                  rows="8"
-                  class="textarea textarea-bordered w-full focus:textarea-primary h-48"
-                  placeholder="You are a helpful AI assistant..."
-                ></textarea>
-              </div>
+        <IntegrationsTab 
+          v-if="activeTab === 'integrations'"
+          :current-settings="currentSettings"
+          :google-auth-status="googleAuthStatus"
+          @connect-google-services="connectGoogleServices"
+          @disconnect-google-services="disconnectGoogleServices"
+        />
 
-              <div>
-                <label
-                  for="tts-voice"
-                  class="block mb-1 text-sm flex items-center"
-                  >Text-to-Speech Voice
-                  <div
-                    class="tooltip tooltip-right"
-                    data-tip="The voice to use for text-to-speech synthesis when Alice speaks responses aloud."
-                  >
-                    <img :src="infoIcon" class="size-4 ml-1" /></div
-                ></label>
-                <select
-                  id="tts-voice"
-                  v-model="currentSettings.ttsVoice"
-                  class="select select-bordered w-full focus:select-primary"
-                >
-                  <option value="alloy">Alloy</option>
-                  <option value="echo">Echo</option>
-                  <option value="fable">Fable</option>
-                  <option value="nova">Nova</option>
-                  <option value="onyx">Onyx</option>
-                  <option value="shimmer">Shimmer</option>
-                </select>
-              </div>
-
-              <!-- GPT-5 Specific Settings -->
-              <div
-                v-if="currentSettings.assistantModel.startsWith('gpt-5')"
-                class="grid grid-cols-1 md:grid-cols-2 gap-4"
-              >
-                <div>
-                  <label
-                    for="reasoning-effort"
-                    class="block mb-1 text-sm flex items-center"
-                    >Reasoning Effort
-                    <div
-                      class="tooltip tooltip-right"
-                      data-tip="Controls how many reasoning tokens the model generates before producing a response. Higher effort provides more thorough reasoning but increases cost and latency."
-                    >
-                      <img :src="infoIcon" class="size-4 ml-1" /></div
-                  ></label>
-                  <select
-                    id="reasoning-effort"
-                    v-model="currentSettings.assistantReasoningEffort"
-                    class="select select-bordered w-full focus:select-primary"
-                  >
-                    <option value="minimal">Minimal (Fastest)</option>
-                    <option value="low">Low</option>
-                    <option value="medium">Medium</option>
-                    <option value="high">High (Most thorough)</option>
-                  </select>
-                </div>
-                <div>
-                  <label
-                    for="response-verbosity"
-                    class="block mb-1 text-sm flex items-center"
-                    >Response Verbosity
-                    <div
-                      class="tooltip tooltip-left"
-                      data-tip="Controls how detailed the model's responses are. Higher verbosity provides more thorough explanations but increases token usage."
-                    >
-                      <img :src="infoIcon" class="size-4 ml-1" /></div
-                  ></label>
-                  <select
-                    id="response-verbosity"
-                    v-model="currentSettings.assistantVerbosity"
-                    class="select select-bordered w-full focus:select-primary"
-                  >
-                    <option value="low">Low (Concise)</option>
-                    <option value="medium">Medium</option>
-                    <option value="high">High (Detailed)</option>
-                  </select>
-                </div>
-              </div>
-
-              <!-- Standard Model Settings (Temperature/Top-P) -->
-              <div
-                v-if="
-                  !currentSettings.assistantModel.startsWith('gpt-5') &&
-                  !currentSettings.assistantModel.startsWith('o')
-                "
-                class="grid grid-cols-1 md:grid-cols-2 gap-4"
-              >
-                <div>
-                  <label
-                    for="assistant-temperature"
-                    class="block mb-1 text-sm flex items-center"
-                  >
-                    Assistant Temperature ({{
-                      currentSettings.assistantTemperature.toFixed(1)
-                    }})
-                    <div
-                      class="tooltip tooltip-right"
-                      data-tip="Controls the randomness of the AI's responses. Higher values (e.g., 0.8) make the output more creative, while lower values (e.g., 0.2) make it more deterministic."
-                    >
-                      <img :src="infoIcon" class="size-4 ml-1" /></div
-                  ></label>
-                  <input
-                    id="assistant-temperature"
-                    type="range"
-                    min="0"
-                    max="2"
-                    step="0.1"
-                    v-model.number="currentSettings.assistantTemperature"
-                    class="range range-primary"
-                  />
-                </div>
-                <div>
-                  <label
-                    for="assistant-top-p"
-                    class="block mb-1 text-sm flex items-center"
-                  >
-                    Assistant Top P ({{
-                      currentSettings.assistantTopP.toFixed(1)
-                    }})
-                    <div
-                      class="tooltip tooltip-left"
-                      data-tip="An alternative to temperature sampling, where the model considers the results of the tokens with top-p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered."
-                    >
-                      <img :src="infoIcon" class="size-4 ml-1" /></div
-                  ></label>
-                  <input
-                    id="assistant-top-p"
-                    type="range"
-                    min="0"
-                    max="1"
-                    step="0.1"
-                    v-model.number="currentSettings.assistantTopP"
-                    class="range range-success"
-                  />
-                </div>
-              </div>
-            </div>
-          </fieldset>
-
-          <fieldset
-            class="fieldset bg-gray-900/90 border-green-500/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">Context & Memory</legend>
-            <div class="space-y-4 p-2">
-              <div class="grid grid-cols-1 md:grid-cols-2 gap-4">
-                <div>
-                  <label
-                    for="max-history-messages"
-                    class="block mb-1 text-sm flex items-center"
-                    >Max History Messages (API)
-                    <div
-                      class="tooltip tooltip-right"
-                      data-tip="The number of recent messages to send to the AI. A higher number provides more context but increases cost."
-                    >
-                      <img :src="infoIcon" class="size-4 ml-1" /></div
-                  ></label>
-                  <input
-                    id="max-history-messages"
-                    type="number"
-                    min="1"
-                    max="50"
-                    step="1"
-                    v-model.number="
-                      currentSettings.MAX_HISTORY_MESSAGES_FOR_API
-                    "
-                    class="input input-bordered w-full focus:input-primary"
-                  />
-                </div>
-                <div>
-                  <label
-                    for="summarization-messages"
-                    class="block mb-1 text-sm flex items-center"
-                    >Summarization Message Count
-                    <div
-                      class="tooltip tooltip-left"
-                      data-tip="The number of messages to summarize for providing context to the AI. This helps the AI remember the conversation over a longer period."
-                    >
-                      <img :src="infoIcon" class="size-4 ml-1" /></div
-                  ></label>
-                  <input
-                    id="summarization-messages"
-                    type="number"
-                    min="5"
-                    max="100"
-                    step="1"
-                    v-model.number="currentSettings.SUMMARIZATION_MESSAGE_COUNT"
-                    class="input input-bordered w-full focus:input-primary"
-                  />
-                </div>
-              </div>
-              <div>
-                <label
-                  for="summarization-model"
-                  class="block mb-1 text-sm flex items-center"
-                  >Summarization Model *
-                  <div
-                    class="tooltip tooltip-right"
-                    data-tip="The AI model used to summarize the conversation. A smaller, faster model is recommended."
-                  >
-                    <img :src="infoIcon" class="size-4 ml-1" /></div
-                ></label>
-                <select
-                  id="summarization-model"
-                  v-model="currentSettings.SUMMARIZATION_MODEL"
-                  class="select select-bordered w-full focus:select-primary"
-                >
-                  <option disabled value="">
-                    Select a summarization model
-                  </option>
-                  <option
-                    v-if="
-                      conversationStore.availableModels.length === 0 &&
-                      settingsStore.coreOpenAISettingsValid
-                    "
-                    value=""
-                  >
-                    Loading models...
-                  </option>
-                  <option
-                    v-for="model in availableModelsForSelect"
-                    :key="`summ-${model.id}`"
-                    :value="model.id"
-                  >
-                    {{ model.id }}
-                  </option>
-                </select>
-                <p
-                  v-if="
-                    !settingsStore.coreOpenAISettingsValid &&
-                    ((currentSettings.aiProvider === 'openai' &&
-                      currentSettings.VITE_OPENAI_API_KEY) ||
-                      (currentSettings.aiProvider === 'openrouter' &&
-                        currentSettings.VITE_OPENROUTER_API_KEY)) &&
-                    conversationStore.availableModels.length === 0
-                  "
-                  class="text-xs text-warning mt-1"
-                >
-                  {{
-                    currentSettings.aiProvider === 'openai'
-                      ? 'OpenAI'
-                      : 'OpenRouter'
-                  }}
-                  API key needs to be validated (Save & Test) to load models.
-                </p>
-                <p class="text-xs text-gray-400 mt-1">
-                  Model used for generating conversation summaries (e.g.,
-                  gpt-4.1-nano).
-                </p>
-              </div>
-
-              <div>
-                <label
-                  for="summarization-system-prompt"
-                  class="block mb-1 text-sm"
-                  >Summarization System Prompt</label
-                >
-                <textarea
-                  id="summarization-system-prompt"
-                  v-model="currentSettings.SUMMARIZATION_SYSTEM_PROMPT"
-                  rows="6"
-                  class="textarea textarea-bordered w-full focus:textarea-primary h-40"
-                  placeholder="You are an expert conversation summarizer..."
-                ></textarea>
-                <p class="text-xs text-gray-400 mt-1">
-                  System prompt to guide the summarization model.
-                </p>
-              </div>
-            </div>
-          </fieldset>
-
-          <fieldset
-            class="fieldset bg-gray-900/90 border-green-500/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">Enabled Tools</legend>
-            <div class="space-y-4 p-2">
-              <div>
-                <label class="block mb-2 text-sm font-medium"
-                  >Enabled Assistant Tools</label
-                >
-                <div
-                  class="space-y-2 p-3 border border-neutral-content/20 rounded-md max-h-60 overflow-y-auto bg-gray-800/50"
-                >
-                  <div
-                    v-if="availableToolsForSelect.length === 0"
-                    class="text-xs text-gray-400"
-                  >
-                    No tools defined.
-                  </div>
-                  <div
-                    v-for="tool in availableToolsForSelect"
-                    :key="tool.name"
-                    class="form-control"
-                  >
-                    <label
-                      class="label cursor-pointer py-1 justify-start gap-3"
-                      :class="{
-                        'opacity-50 cursor-not-allowed': !isToolConfigured(
-                          tool.name
-                        ),
-                      }"
-                    >
-                      <input
-                        type="checkbox"
-                        :value="tool.name"
-                        v-model="currentSettings.assistantTools"
-                        class="checkbox checkbox-accent checkbox-sm"
-                        :disabled="!isToolConfigured(tool.name)"
-                      />
-                      <span
-                        class="label-text"
-                        :title="
-                          tool.description +
-                          (!isToolConfigured(tool.name)
-                            ? ' (API key for this tool not configured in Optional Tool APIs section)'
-                            : '')
-                        "
-                      >
-                        {{ tool.displayName }}
-                        <span
-                          v-if="!isToolConfigured(tool.name)"
-                          class="text-xs text-warning normal-case"
-                        >
-                          (Needs config below)
-                        </span>
-                      </span>
-                    </label>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </fieldset>
-
-          <fieldset
-            v-if="isBrowserContextToolActive"
-            class="fieldset bg-gray-900/90 border-green-500/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">WebSocket Configuration</legend>
-            <div class="space-y-4 p-2">
-              <div>
-                <label
-                  for="websocket-port"
-                  class="block mb-1 text-sm flex items-center"
-                >
-                  WebSocket Port
-                  <div
-                    class="tooltip tooltip-right"
-                    data-tip="The port number for the WebSocket server used by the browser_context tool. Ensure this port is available and not blocked by firewall."
-                  >
-                    <img :src="infoIcon" class="size-4 ml-1" />
-                  </div>
-                </label>
-                <input
-                  id="websocket-port"
-                  type="number"
-                  min="1"
-                  max="65535"
-                  step="1"
-                  v-model.number="currentSettings.websocketPort"
-                  class="input input-bordered w-full focus:input-primary"
-                />
-                <p class="text-xs text-gray-400 mt-1">
-                  Port number for WebSocket server (1-65535). Default: 5421
-                </p>
-              </div>
-            </div>
-          </fieldset>
-        </div>
-
-        <div v-if="activeTab === 'hotkeys'" class="space-y-6">
-          <h3 class="text-xl font-semibold mb-4 text-yellow-400">
-            Global Hotkeys
-          </h3>
-          <fieldset
-            class="fieldset bg-gray-900/90 border-yellow-500/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">Keyboard Shortcuts</legend>
-            <div class="p-2 space-y-4">
-              <div>
-                <label for="mic-toggle-hotkey" class="block mb-1 text-sm"
-                  >Microphone Toggle Hotkey</label
-                >
-                <div class="flex items-center justify-between">
-                  <kbd class="kbd kbd-xl">{{
-                    formatAccelerator(currentSettings.microphoneToggleHotkey)
-                  }}</kbd>
-                  <div class="flex items-center gap-2">
-                    <button
-                      type="button"
-                      @click="startRecordingHotkey('microphoneToggleHotkey')"
-                      class="btn btn-secondary btn-active btn-sm"
-                      :disabled="
-                        isRecordingHotkeyFor === 'microphoneToggleHotkey'
-                      "
-                    >
-                      {{
-                        isRecordingHotkeyFor === 'microphoneToggleHotkey'
-                          ? 'Recording...'
-                          : 'Record'
-                      }}
-                    </button>
-                    <button
-                      type="button"
-                      @click="clearHotkey('microphoneToggleHotkey')"
-                      class="btn btn-warning btn-outline btn-sm"
-                      :disabled="!currentSettings.microphoneToggleHotkey"
-                    >
-                      Clear
-                    </button>
-                  </div>
-                </div>
-                <p
-                  v-if="isRecordingHotkeyFor === 'microphoneToggleHotkey'"
-                  class="text-xs text-yellow-400 mt-1"
-                >
-                  Press the desired key combination. Press Esc to cancel.
-                </p>
-              </div>
-
-              <div>
-                <label for="mute-playback-hotkey" class="block mb-1 text-sm"
-                  >Mute Playback Hotkey</label
-                >
-                <div class="flex items-center justify-between">
-                  <kbd class="kbd kbd-xl">{{
-                    formatAccelerator(currentSettings.mutePlaybackHotkey)
-                  }}</kbd>
-                  <div class="flex items-center gap-2">
-                    <button
-                      type="button"
-                      @click="startRecordingHotkey('mutePlaybackHotkey')"
-                      class="btn btn-secondary btn-active btn-sm"
-                      :disabled="isRecordingHotkeyFor === 'mutePlaybackHotkey'"
-                    >
-                      {{
-                        isRecordingHotkeyFor === 'mutePlaybackHotkey'
-                          ? 'Recording...'
-                          : 'Record'
-                      }}
-                    </button>
-                    <button
-                      type="button"
-                      @click="clearHotkey('mutePlaybackHotkey')"
-                      class="btn btn-warning btn-outline btn-sm"
-                      :disabled="!currentSettings.mutePlaybackHotkey"
-                    >
-                      Clear
-                    </button>
-                  </div>
-                </div>
-                <p
-                  v-if="isRecordingHotkeyFor === 'mutePlaybackHotkey'"
-                  class="text-xs text-yellow-400 mt-1"
-                >
-                  Press the desired key combination. Press Esc to cancel.
-                </p>
-              </div>
-
-              <div>
-                <label for="take-screenshot-hotkey" class="block mb-1 text-sm"
-                  >Take Screenshot Hotkey</label
-                >
-                <div class="flex items-center justify-between">
-                  <kbd class="kbd kbd-xl">{{
-                    formatAccelerator(currentSettings.takeScreenshotHotkey)
-                  }}</kbd>
-                  <div class="flex items-center gap-2">
-                    <button
-                      type="button"
-                      @click="startRecordingHotkey('takeScreenshotHotkey')"
-                      class="btn btn-secondary btn-active btn-sm"
-                      :disabled="
-                        isRecordingHotkeyFor === 'takeScreenshotHotkey'
-                      "
-                    >
-                      {{
-                        isRecordingHotkeyFor === 'takeScreenshotHotkey'
-                          ? 'Recording...'
-                          : 'Record'
-                      }}
-                    </button>
-                    <button
-                      type="button"
-                      @click="clearHotkey('takeScreenshotHotkey')"
-                      class="btn btn-warning btn-outline btn-sm"
-                      :disabled="!currentSettings.takeScreenshotHotkey"
-                    >
-                      Clear
-                    </button>
-                  </div>
-                </div>
-                <p
-                  v-if="isRecordingHotkeyFor === 'takeScreenshotHotkey'"
-                  class="text-xs text-yellow-400 mt-1"
-                >
-                  Press the desired key combination. Press Esc to cancel.
-                </p>
-              </div>
-            </div>
-          </fieldset>
-        </div>
-
-        <div v-if="activeTab === 'integrations'" class="space-y-6">
-          <h3 class="text-xl font-semibold mb-4 text-purple-400">
-            Integrations & APIs
-          </h3>
-
-          <fieldset
-            class="fieldset bg-gray-900/90 border-purple-500/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">Google Services Integration</legend>
-            <div class="p-2 space-y-4">
-              <div
-                v-if="
-                  !googleAuthStatus.isAuthenticated &&
-                  !googleAuthStatus.authInProgress
-                "
-              >
-                <button
-                  type="button"
-                  @click="connectGoogleServices"
-                  class="btn btn-info btn-active"
-                  :disabled="googleAuthStatus.isLoading"
-                >
-                  {{
-                    googleAuthStatus.isLoading
-                      ? 'Connecting...'
-                      : 'Connect to Google Services'
-                  }}
-                </button>
-              </div>
-              <div
-                v-if="
-                  googleAuthStatus.authInProgress &&
-                  !googleAuthStatus.isAuthenticated
-                "
-              >
-                <p class="text-sm mb-2">
-                  Awaiting authorization in your browser. Please follow the
-                  instructions there.
-                </p>
-                <span class="loading loading-dots loading-md"></span>
-              </div>
-              <div v-if="googleAuthStatus.isAuthenticated">
-                <div role="alert" class="alert alert-success mb-4">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    class="h-6 w-6 shrink-0 stroke-current"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      stroke-linecap="round"
-                      stroke-linejoin="round"
-                      stroke-width="2"
-                      d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"
-                    />
-                  </svg>
-                  <span>Successfully connected to Google Services.</span>
-                </div>
-                <button
-                  type="button"
-                  @click="disconnectGoogleServices"
-                  class="btn btn-warning btn-outline"
-                >
-                  Disconnect from Google Services
-                </button>
-              </div>
-              <p
-                v-if="googleAuthStatus.error"
-                class="text-xs text-error-content mt-1"
-              >
-                {{ googleAuthStatus.error }}
-              </p>
-              <p
-                v-if="
-                  googleAuthStatus.message &&
-                  !googleAuthStatus.isAuthenticated &&
-                  !googleAuthStatus.error
-                "
-                class="text-xs text-white mt-1"
-              >
-                {{ googleAuthStatus.message }}
-              </p>
-            </div>
-          </fieldset>
-
-          <fieldset
-            class="fieldset bg-gray-900/90 border-cyan-500/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">Remote MCP Servers</legend>
-            <div class="p-2 space-y-4">
-              <div>
-                <label for="mcp-servers-config" class="block mb-1 text-sm">
-                  MCP Servers JSON Configuration (Array)
-                </label>
-                <textarea
-                  id="mcp-servers-config"
-                  v-model="currentSettings.mcpServersConfig"
-                  rows="10"
-                  class="textarea textarea-bordered w-full focus:textarea-primary h-60 bg-gray-800"
-                  :placeholder="mcpPlaceholder"
-                ></textarea>
-                <p class="text-xs text-gray-400 mt-1">
-                  Enter a JSON array of MCP server configurations. Each object
-                  should follow the
-                  <a
-                    href="https://cookbook.openai.com/examples/mcp/mcp_tool_guide"
-                    target="_blank"
-                    class="link link-hover"
-                    >OpenAI MCP tool format</a
-                  >.
-                </p>
-              </div>
-            </div>
-          </fieldset>
-
-          <fieldset
-            class="fieldset bg-gray-900/90 border-gray-600/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">
-              Optional Tool APIs
-              <a
-                href="https://github.com/pmbstyle/Alice/blob/main/docs/toolsInstructions.md"
-                target="_blank"
-                class="ml-2"
-              >
-                <span class="badge badge-sm badge-soft whitespace-nowrap">
-                  Info
-                  <img :src="newTabIcon" class="size-3 inline-block ml-1" />
-                </span>
-              </a>
-            </legend>
-            <div class="grid grid-cols-1 md:grid-cols-2 gap-4 p-2">
-              <div>
-                <label for="jackett-url" class="block mb-1 text-sm"
-                  >Jackett URL (Torrent Search)</label
-                >
-                <input
-                  id="jackett-url"
-                  type="text"
-                  v-model="currentSettings.VITE_JACKETT_URL"
-                  class="input focus:outline-none w-full"
-                />
-              </div>
-              <div>
-                <label for="jackett-key" class="block mb-1 text-sm"
-                  >Jackett API Key</label
-                >
-                <input
-                  id="jackett-key"
-                  type="password"
-                  v-model="currentSettings.VITE_JACKETT_API_KEY"
-                  class="input focus:outline-none w-full"
-                  autocomplete="new-password"
-                />
-              </div>
-              <div>
-                <label for="qb-url" class="block mb-1 text-sm"
-                  >qBittorrent URL</label
-                >
-                <input
-                  id="qb-url"
-                  type="text"
-                  v-model="currentSettings.VITE_QB_URL"
-                  class="input focus:outline-none w-full"
-                />
-              </div>
-              <div>
-                <label for="qb-user" class="block mb-1 text-sm"
-                  >qBittorrent Username</label
-                >
-                <input
-                  id="qb-user"
-                  type="text"
-                  v-model="currentSettings.VITE_QB_USERNAME"
-                  class="input focus:outline-none w-full"
-                />
-              </div>
-              <div>
-                <label for="qb-pass" class="block mb-1 text-sm"
-                  >qBittorrent Password</label
-                >
-                <input
-                  id="qb-pass"
-                  type="password"
-                  v-model="currentSettings.VITE_QB_PASSWORD"
-                  class="input focus:outline-none w-full"
-                  autocomplete="new-password"
-                />
-              </div>
-              <div
-                v-if="currentSettings.assistantTools.includes('perform_web_search')"
-              >
-                <label for="tavily-key" class="block mb-1 text-sm"
-                  >Tavily API Key (Web Search)</label
-                >
-                <input
-                  id="tavily-key"
-                  type="password"
-                  v-model="currentSettings.VITE_TAVILY_API_KEY"
-                  class="input focus:outline-none w-full"
-                  autocomplete="new-password"
-                  placeholder="tvly-..."
-                />
-                <p class="text-xs text-gray-400 mt-1">
-                  Required for web search functionality. Get your API key from 
-                  <a 
-                    href="https://tavily.com" 
-                    target="_blank" 
-                    class="link link-primary"
-                  >Tavily</a>.
-                </p>
-              </div>
-            </div>
-          </fieldset>
-        </div>
-
-        <div v-if="activeTab === 'security'" class="space-y-6">
-          <h3 class="text-xl font-semibold mb-4 text-red-400">
-            Security & Command Permissions
-          </h3>
-
-          <fieldset
-            class="fieldset bg-gray-900/90 border-red-500/50 rounded-box w-full border p-4"
-          >
-            <legend class="fieldset-legend">Approved Commands</legend>
-            <div class="p-2 space-y-4">
-              <div class="text-sm text-gray-300 mb-4">
-                These commands can be executed by Alice without requiring
-                approval. Commands approved "for session" are shown in the
-                current session column.
-              </div>
-
-              <div class="overflow-x-auto">
-                <table class="table table-zebra table-sm w-full">
-                  <thead>
-                    <tr>
-                      <th>Command</th>
-                      <th>Approval Type</th>
-                      <th>Actions</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <tr
-                      v-for="command in settingsStore.settings.approvedCommands"
-                      :key="command"
-                    >
-                      <td class="font-mono text-sm">{{ command }}</td>
-                      <td>
-                        <span class="badge badge-success badge-sm"
-                          >Permanent</span
-                        >
-                      </td>
-                      <td>
-                        <button
-                          @click="removeCommand(command)"
-                          class="btn btn-error btn-xs"
-                          title="Remove command"
-                        >
-                          ✗
-                        </button>
-                      </td>
-                    </tr>
-                    <tr
-                      v-for="command in settingsStore.sessionApprovedCommands"
-                      :key="'session-' + command"
-                    >
-                      <td class="font-mono text-sm">{{ command }}</td>
-                      <td>
-                        <span class="badge badge-info badge-sm">Session</span>
-                      </td>
-                      <td>
-                        <span class="text-xs text-gray-500"
-                          >Auto-removed on restart</span
-                        >
-                      </td>
-                    </tr>
-                  </tbody>
-                </table>
-              </div>
-
-              <div
-                v-if="
-                  settingsStore.settings.approvedCommands.length === 0 &&
-                  settingsStore.sessionApprovedCommands.length === 0
-                "
-                class="text-center py-4 text-gray-400"
-              >
-                No approved commands. Commands will require approval before
-                execution.
-              </div>
-            </div>
-          </fieldset>
-        </div>
+        <SecurityTab 
+          v-if="activeTab === 'security'"
+          :approved-commands="settingsStore.settings.approvedCommands"
+          :session-approved-commands="settingsStore.sessionApprovedCommands"
+          @remove-command="removeCommand"
+        />
       </div>
 
       <div class="mt-8 flex flex-col sm:flex-row justify-center gap-4">
@@ -1154,13 +186,20 @@
 </template>
 
 <script setup lang="ts">
-import { ref, watch, onMounted, onUnmounted, reactive, computed } from 'vue'
+import { ref, watch, onMounted, computed } from 'vue'
 import { storeToRefs } from 'pinia'
 import { useSettingsStore, type AliceSettings } from '../stores/settingsStore'
 import { useConversationStore } from '../stores/conversationStore'
-import { newTabIcon, heartIcon, infoIcon } from '../utils/assetsImport'
+import { heartIcon } from '../utils/assetsImport'
 import { PREDEFINED_OPENAI_TOOLS } from '../utils/assistantTools'
 import { DEFAULT_ASSISTANT_SYSTEM_PROMPT } from '../stores/settingsStore'
+import { useHotkeyRecording } from '../composables/useHotkeyRecording'
+import { useGoogleAuth } from '../composables/useGoogleAuth'
+import CoreSettingsTab from './settings/CoreSettingsTab.vue'
+import AssistantSettingsTab from './settings/AssistantSettingsTab.vue'
+import HotkeysTab from './settings/HotkeysTab.vue'
+import IntegrationsTab from './settings/IntegrationsTab.vue'
+import SecurityTab from './settings/SecurityTab.vue'
 
 const appVersion = ref(import.meta.env.VITE_APP_VERSION || '')
 const settingsStore = useSettingsStore()
@@ -1171,19 +210,21 @@ const activeTab = ref<
   'core' | 'assistant' | 'hotkeys' | 'integrations' | 'security'
 >('core')
 
-const isRecordingHotkeyFor = ref<keyof AliceSettings | null>(null)
-const activeRecordingKeys = ref<Set<string>>(new Set())
 const isRefreshingModels = ref(false)
 
 const { availableModels } = storeToRefs(conversationStore)
 
-const googleAuthStatus = reactive({
-  isAuthenticated: false,
-  authInProgress: false,
-  isLoading: false,
-  error: null as string | null,
-  message: null as string | null,
-})
+const {
+  isRecordingHotkeyFor,
+  startRecordingHotkey: startRecordingHotkeyComposable,
+  clearHotkey: clearHotkeyComposable,
+} = useHotkeyRecording()
+
+const {
+  googleAuthStatus,
+  connectGoogleServices,
+  disconnectGoogleServices,
+} = useGoogleAuth()
 
 const availableToolsForSelect = computed(() => {
   return PREDEFINED_OPENAI_TOOLS.map(tool => {
@@ -1229,16 +270,6 @@ const refreshModels = async () => {
   }
 }
 
-const getProviderDisplayName = (provider: string): string => {
-  const providerNames: Record<string, string> = {
-    'openai': 'OpenAI',
-    'openrouter': 'OpenRouter',
-    'ollama': 'Ollama',
-    'lm-studio': 'LM Studio'
-  }
-  return providerNames[provider] || provider
-}
-
 function betterToolName(name: string): string {
   const nameMap: Record<string, string> = {
     get_current_datetime: 'Current Date & Time',
@@ -1264,18 +295,16 @@ function betterToolName(name: string): string {
   )
 }
 
-const mcpPlaceholder = `[
-// MCP Servers array, example server:
-  {
-    "type": "mcp",
-    "server_label": "deepwiki",
-    "server_url": "https://mcp.deepwiki.com/mcp",
-    "require_approval": "never"
-  }
-]`
-
 const resetSystemPrompt = () => {
   currentSettings.value.assistantSystemPrompt = DEFAULT_ASSISTANT_SYSTEM_PROMPT
+}
+
+const startRecordingHotkey = (settingKey: keyof AliceSettings) => {
+  startRecordingHotkeyComposable(settingKey, currentSettings.value)
+}
+
+const clearHotkey = (settingKey: keyof AliceSettings) => {
+  clearHotkeyComposable(settingKey, currentSettings.value)
 }
 
 function isToolConfigured(toolName: string): boolean {
@@ -1295,194 +324,6 @@ function isToolConfigured(toolName: string): boolean {
   })
 }
 
-async function checkGoogleAuthStatus() {
-  googleAuthStatus.isLoading = true
-  googleAuthStatus.error = null
-  try {
-    const result = await window.ipcRenderer.invoke(
-      'google-calendar:check-auth-status'
-    )
-    if (result.success)
-      googleAuthStatus.isAuthenticated = result.isAuthenticated
-  } catch (e: any) {
-    googleAuthStatus.error = 'Error checking auth status: ' + e.message
-  } finally {
-    googleAuthStatus.isLoading = false
-  }
-}
-
-async function connectGoogleServices() {
-  googleAuthStatus.isLoading = true
-  googleAuthStatus.authInProgress = true
-  googleAuthStatus.error = null
-  googleAuthStatus.message = null
-  try {
-    const result = await window.ipcRenderer.invoke(
-      'google-calendar:get-auth-url'
-    )
-    if (result.success) googleAuthStatus.message = result.message
-    else {
-      googleAuthStatus.error =
-        result.error || 'Failed to start Google authentication.'
-      googleAuthStatus.authInProgress = false
-    }
-  } catch (e: any) {
-    googleAuthStatus.error = 'Error initiating auth: ' + e.message
-    googleAuthStatus.authInProgress = false
-  } finally {
-    googleAuthStatus.isLoading = false
-  }
-}
-
-async function disconnectGoogleServices() {
-  googleAuthStatus.isLoading = true
-  googleAuthStatus.error = null
-  googleAuthStatus.message = 'Disconnecting...'
-  try {
-    const result = await window.ipcRenderer.invoke('google-calendar:disconnect')
-    if (result.success) {
-      googleAuthStatus.isAuthenticated = false
-      googleAuthStatus.authInProgress = false
-      googleAuthStatus.message = result.message
-    } else {
-      googleAuthStatus.error =
-        result.error || 'Failed to disconnect from Google.'
-      googleAuthStatus.message = null
-    }
-  } catch (e: any) {
-    googleAuthStatus.error = 'Error disconnecting: ' + e.message
-    googleAuthStatus.message = null
-  } finally {
-    googleAuthStatus.isLoading = false
-  }
-}
-
-function handleGoogleAuthSuccess(event: any, message: string) {
-  googleAuthStatus.isAuthenticated = true
-  googleAuthStatus.authInProgress = false
-  googleAuthStatus.message = message
-  googleAuthStatus.error = null
-}
-
-function handleGoogleAuthError(event: any, errorMsg: string) {
-  googleAuthStatus.isAuthenticated = false
-  googleAuthStatus.authInProgress = false
-  googleAuthStatus.error = `Authentication failed: ${errorMsg}`
-  googleAuthStatus.message = null
-}
-
-const modifierKeys = [
-  'Control',
-  'Alt',
-  'Shift',
-  'Meta',
-  'Command',
-  'Cmd',
-  'Option',
-  'Super',
-]
-const keyToAcceleratorMap: Record<string, string> = {
-  Control: 'Control',
-  Alt: 'Alt',
-  Shift: 'Shift',
-  Meta: 'Super',
-  ArrowUp: 'Up',
-  ArrowDown: 'Down',
-  ArrowLeft: 'Left',
-  ArrowRight: 'Right',
-  Escape: 'Esc',
-  Enter: 'Return',
-  Delete: 'Delete',
-  Backspace: 'Backspace',
-  Tab: 'Tab',
-  PageUp: 'PageUp',
-  PageDown: 'PageDown',
-  Home: 'Home',
-  End: 'End',
-  Insert: 'Insert',
-  Space: 'Space',
-  '+': 'Plus',
-}
-
-function electronAcceleratorForKey(key: string): string {
-  if (key.length === 1 && key.match(/[a-z0-9]/i)) {
-    return key.toUpperCase()
-  }
-  return keyToAcceleratorMap[key] || key
-}
-
-function formatAccelerator(accelerator: string | undefined): string {
-  if (!accelerator) return ''
-  return accelerator.replace(/\+/g, ' + ')
-}
-
-const handleHotkeyKeyDown = (event: KeyboardEvent) => {
-  if (!isRecordingHotkeyFor.value) return
-
-  event.preventDefault()
-  event.stopPropagation()
-
-  const key = event.key
-  if (key === 'Escape') {
-    stopRecordingHotkey()
-    return
-  }
-
-  if (
-    modifierKeys.includes(key) ||
-    modifierKeys.map(m => m.toLowerCase()).includes(key.toLowerCase())
-  ) {
-    activeRecordingKeys.value.add(
-      electronAcceleratorForKey(
-        key === 'Meta' && navigator.platform.toUpperCase().indexOf('MAC') >= 0
-          ? 'Command'
-          : key === 'Meta'
-            ? 'Super'
-            : key
-      )
-    )
-  } else {
-    const mainKey = electronAcceleratorForKey(key)
-    activeRecordingKeys.value.add(mainKey)
-    const acceleratorParts = Array.from(activeRecordingKeys.value)
-
-    const order = ['Control', 'Alt', 'Shift', 'Super', 'Command']
-    acceleratorParts.sort((a, b) => {
-      const aIsModifier = order.includes(a)
-      const bIsModifier = order.includes(b)
-      if (aIsModifier && !bIsModifier) return -1
-      if (!aIsModifier && bIsModifier) return 1
-      if (aIsModifier && bIsModifier) return order.indexOf(a) - order.indexOf(b)
-      return a.localeCompare(b)
-    })
-
-    const newHotkey = acceleratorParts.join('+')
-    if (isRecordingHotkeyFor.value) {
-      (currentSettings.value[isRecordingHotkeyFor.value as keyof AliceSettings] as any) =
-        newHotkey
-    }
-    stopRecordingHotkey()
-  }
-}
-
-const startRecordingHotkey = (settingKey: keyof AliceSettings) => {
-  isRecordingHotkeyFor.value = settingKey
-  activeRecordingKeys.value.clear()
-  window.addEventListener('keydown', handleHotkeyKeyDown, true)
-}
-
-const stopRecordingHotkey = () => {
-  window.removeEventListener('keydown', handleHotkeyKeyDown, true)
-  isRecordingHotkeyFor.value = null
-  activeRecordingKeys.value.clear()
-}
-
-const clearHotkey = (settingKey: keyof AliceSettings) => {
-  if (isRecordingHotkeyFor.value === settingKey) {
-    stopRecordingHotkey()
-  }
-  ;(currentSettings.value[settingKey] as any) = ''
-}
 
 onMounted(async () => {
   if (!settingsStore.initialLoadAttempted) {
@@ -1495,28 +336,6 @@ onMounted(async () => {
     conversationStore.availableModels.length === 0
   ) {
     await conversationStore.fetchModels()
-  }
-
-  await checkGoogleAuthStatus()
-  if (window.ipcRenderer) {
-    window.ipcRenderer.on(
-      'google-auth-loopback-success',
-      handleGoogleAuthSuccess
-    )
-    window.ipcRenderer.on('google-auth-loopback-error', handleGoogleAuthError)
-  }
-})
-
-onUnmounted(() => {
-  if (window.ipcRenderer) {
-    window.ipcRenderer.off(
-      'google-auth-loopback-success',
-      handleGoogleAuthSuccess
-    )
-    window.ipcRenderer.off('google-auth-loopback-error', handleGoogleAuthError)
-  }
-  if (isRecordingHotkeyFor.value) {
-    stopRecordingHotkey()
   }
 })
 
@@ -1550,9 +369,6 @@ watch(
 )
 
 const handleSaveAndTestSettings = async () => {
-  if (isRecordingHotkeyFor.value) {
-    stopRecordingHotkey()
-  }
   if (
     currentSettings.value.mcpServersConfig &&
     currentSettings.value.mcpServersConfig.trim() !== '' &&
